% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animals1}
\alias{animals1}
\title{Simulated preliminary animals uncensored data}
\format{
This data frame contains the following columns:

  \describe{

  \item{\code{ID}}{PDX line identifier.}

  \item{\code{Y}}{time-to-event variable.}

  \item{\code{Tx}}{treatment indicator. \code{0} denotes the placebo group and \code{1} the treatment group.}

  }
}
\usage{
data(animals1)
}
\description{
The \code{animals1} data frame has 18 rows and 3 columns, with all uncensored observations.
}
\keyword{datasets}
