% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootfit.R
\name{bootfit}
\alias{bootfit}
\title{A function to calculate the power under pre-specified effect size,
variance, and correlation using Monte Carlo sampling scheme by fitting ANOVA fixed effects model,
ANOVA random effects model, Cox model, and frailty model.}
\usage{
bootfit(
  seed = 1000,
  n,
  m,
  beta,
  tau2,
  alpha = 0.05,
  lambda = 0.03,
  nu = 2,
  sigma2 = 1,
  distr = c("Weibull", "normal"),
  two.sided = TRUE,
  lambdaC = 0.1,
  censor = TRUE,
  model = c("both", "ANOVA", "Cox-frailty")
)
}
\arguments{
\item{seed}{an integer random seed number.}

\item{n}{an integer number to specify the number of PDX lines.}

\item{m}{an integer number to specify the number of individuals per PDX line per treatment.}

\item{beta}{Treatment effect for the treated group.}

\item{tau2}{Error variance of random effect.}

\item{alpha}{Significance level. Default is 0.05.}

\item{lambda}{Scale parameter of Weibull distribution for the baseline hazard.}

\item{nu}{Shape parameter of Weibull distribution for the baseline hazard.}

\item{sigma2}{Error variance of log survival time for both treatment groups.}

\item{distr}{distributional assumption of survival time.}

\item{two.sided}{Logical value to indicate if a two-sided hypothesis testing is conducted.}

\item{lambdaC}{Rate parameter of exponential distribution for the hazard of censoring.}

\item{censor}{logical value of whether a censoring distribution is considered in a data generation setting. Default is TRUE.}

\item{model}{a string to indicate which model is fitted.}
}
\description{
A function to calculate the power under pre-specified effect size, variance, and
correlation using Monte Carlo sampling scheme
}
