% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpower.R
\name{plotpower}
\alias{plotpower}
\title{A function to generate a four-panel power curve under specified
number of PDX lines and number of individuals per PDX lines per treatment}
\usage{
plotpower(
  object,
  ylim = c(0, 0.1),
  x.axis.n.breaks = NULL,
  y.axis.n.breaks = NULL,
  cut.off = 0
)
}
\arguments{
\item{object}{object of class 'PowerTable'.}

\item{ylim}{limit of y axis.}

\item{x.axis.n.breaks}{number of breaks of the x axis. Default is NULL.}

\item{y.axis.n.breaks}{number of breaks of the y axis. Default is NULL.}

\item{cut.off}{a number to indicate the cut-off for power analysis. Default is 0.}
}
\value{
a figure generated by ggplot.
}
\description{
A function to generate a four-panel power curve under specified
number of PDX lines and number of individuals per PDX lines per treatment
}
