% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerTable.R
\name{PowerTable}
\alias{PowerTable}
\title{A function to obtain a power table with the combination of
various number of PDX lines and number of individuals per PDX lines per treatment.}
\usage{
PowerTable(
  n,
  m,
  beta,
  tau2 = 0.5,
  alpha = 0.05,
  lambda = 0.03,
  nu = 2,
  sigma2 = 1,
  two.sided = TRUE,
  distr = c("weibull", "normal"),
  lambdaC = 0.1,
  censor = TRUE,
  sim = 1000,
  ncores = 10
)
}
\arguments{
\item{n}{an integer number to specify the number of PDX lines.}

\item{m}{an integer number to specify the number of individuals per PDX line per treatment.}

\item{beta}{Treatment effect for the treated group.}

\item{alpha}{Signifiance level. Default is 0.05.}

\item{lambda}{Scale parameter of weibull distribution for the baseline hazard.}

\item{nu}{Shape parameter of weibull distribution for the baseline hazard.}

\item{sigma2}{Error variance of log survival time for both treatment groups.}

\item{two.sided}{A logical value to indicate if a two-sided hypothesis testing is conducted.
Default is TRUE.}

\item{distr}{Distributional assumption of the simulated event time.}

\item{lambdaC}{Rate parameter of exponential distribution for the hazard of censoring.}

\item{censor}{logical value of whether a censoring distribution is considered in a data generation setting. Deafult is TRUE.}

\item{sim}{Number of Monte Carlo samples to be generated. Default is 1000.}

\item{ncores}{number of cores for parallel computation.}
}
\description{
A function to obtain a power table with the combination of
various number of PDX lines and number of individuals per PDX lines per treatment.
}
\examples{

require(PDXpower)
n = c(4, 6)
m = c(1, 2)
a <- PowerTable(n = n, m = m, betaB = 0,
                tau2 = 0.5, sim = 10)
plotpower(a)
}
\seealso{
\code{\link{plotpower}}
}
