% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PowANOVADat.R
\name{print.PowANOVADat}
\alias{print.PowANOVADat}
\title{Print PowANOVADat}
\usage{
\method{print}{PowANOVADat}(x, digits = 4, ...)
}
\arguments{
\item{x}{object of class 'PowANOVADat'.}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a summary of power analysis including parameter estimates and statistical power.
}
\description{
Print PowANOVADat
}
\seealso{
\code{\link{PowANOVADat}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
