% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PowerTable.R
\name{print.PowerTable}
\alias{print.PowerTable}
\title{Print PowerTable}
\usage{
\method{print}{PowerTable}(x, digits = 2, ...)
}
\arguments{
\item{x}{object of class 'PowerTable'.}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Object of \code{print.PowerTable} with elements
\item{data}{a data frame of estimated power across all combinations and \code{n} and \code{m}.}
}
\description{
Print PowerTable
}
\seealso{
\code{\link{PowerTable}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
