% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PowFrailtyDat.R
\name{print.PowFrailtyDat}
\alias{print.PowFrailtyDat}
\title{Print PowFrailtyDat}
\usage{
\method{print}{PowFrailtyDat}(x, digits = 4, ...)
}
\arguments{
\item{x}{object of class 'PowFrailtyDat'.}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
a summary of power analysis including parameter estimates and statistical power.
}
\description{
Print PowFrailtyDat
}
\seealso{
\code{\link{PowFrailtyDat}}
}
\author{
Shanpeng Li \email{lishanpeng0913@ucla.edu}
}
