% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPDXdata.R
\name{SimPDXdata}
\alias{SimPDXdata}
\title{A function to calculate the power under pre-specified effect size,
variance, and correlation using Monte Carlo sampling scheme by fitting ANCOVA fixed effects model,
ANCOVA random effects model, Cox model, and frailty model.}
\usage{
SimPDXdata(
  seed = 1000,
  n,
  m,
  beta,
  tau2,
  lambda = 0.03,
  nu = 2,
  sigma2 = 1,
  distr = c("weibull", "normal"),
  two.sided = TRUE,
  lambdaC = 0.1,
  censor = TRUE
)
}
\arguments{
\item{seed}{an integer random seed number.}

\item{n}{an integer number to specify the number of PDX lines.}

\item{m}{an integer number to specify the number of individuals per PDX line per treatment.}

\item{beta}{Treatment effect for the treated group.}

\item{tau2}{Error variance of random effect.}

\item{lambda}{Scale parameter of weibull distribution for the baseline hazard.}

\item{nu}{Shape parameter of weibull distribution for the baseline hazard.}

\item{sigma2}{Error variance of log survival time for both treatment groups.}

\item{distr}{distributional assumption of survival time.}

\item{lambdaC}{Scale parameter of exponential distribution for the hazard of censoring.}

\item{censor}{logical value of whether a censoring distribution is considered in a data generation setting. Deafult is TRUE.}
}
\description{
A function to calculate the power under pre-specified effect size, variance, and
correlation using Monte Carlo sampling scheme
}
