% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpower.R
\name{plotpower}
\alias{plotpower}
\title{A function to generate a four-panel power curve under specified
number of PDX lines and number of individuals per PDX lines per treatment}
\usage{
plotpower(
  object,
  ylim = c(0, 0.1),
  x.axis.n.breaks = 10,
  y.axis.n.breaks = 10,
  cut.off = 0.8,
  legend.position = "right"
)
}
\arguments{
\item{object}{object of class 'PowerTable'.}

\item{ylim}{limit of y axis.}

\item{x.axis.n.breaks}{number of breaks of the x axis. Default is 10.}

\item{y.axis.n.breaks}{number of breaks of the y axis. Default is 10.}

\item{cut.off}{a number to indicate the cut-off for power analysis. Default is 0.8.}

\item{legend.position}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector).}
}
\value{
a figure generated by ggplot.
}
\description{
A function to generate a four-panel power curve under specified
number of PDX lines and number of individuals per PDX lines per treatment
}
