% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpower.R
\name{plotpower}
\alias{plotpower}
\title{A functon to generate a four-panel power curve under specified
number of PDX lines and number of individuals per PDX lines per treatment}
\usage{
plotpower(object, ylim = c(0, 0.1))
}
\arguments{
\item{ylim}{The limit of y axis.}

\item{objectObject}{of class 'PowerTable'.}
}
\description{
A functon to generate a four-panel power curve under specified
number of PDX lines and number of individuals per PDX lines per treatment
}
