% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpower.R
\name{plotpower}
\alias{plotpower}
\title{A function to generate a four-panel power curve under specified
number of PDX lines and number of individuals per PDX lines per treatment}
\usage{
plotpower(object, ylim = c(0, 0.1), n.breaks = NULL)
}
\arguments{
\item{object}{object of class 'PowerTable'.}

\item{ylim}{limit of y axis.}

\item{n.breaks}{number of breaks of the x axis. Default is NULL.}
}
\description{
A function to generate a four-panel power curve under specified
number of PDX lines and number of individuals per PDX lines per treatment
}
