% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpower.R
\name{plotpower}
\alias{plotpower}
\title{A function to generate a four-panel power curve under specified
number of PDX lines and number of individuals per PDX lines per treatment}
\usage{
plotpower(object, ylim = c(0, 0.1))
}
\arguments{
\item{object}{of class 'PowerTable'.}

\item{ylim}{The limit of y axis.}
}
\description{
A function to generate a four-panel power curve under specified
number of PDX lines and number of individuals per PDX lines per treatment
}
