% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowANOVADat.R
\name{PowANOVADat}
\alias{PowANOVADat}
\title{A function to obtain a power table with the combination of
various number of PDX lines and number of individuals per PDX lines per treatment based on a preliminary dataset}
\usage{
PowANOVADat(
  data = NULL,
  formula = NULL,
  random = NULL,
  n = NULL,
  m = NULL,
  sim = 100,
  two.sided = TRUE,
  ncores = NULL
)
}
\arguments{
\item{data}{data.frame in which to interpret the variables named in the formula.}

\item{formula}{a two-sided linear formula object describing the fixed-effects part of the model,
with the response on the left of a ~ operator and the terms, separated by + operators, on the right.}

\item{random}{an one-sided formula of the form \code{~ x1 + ... + xn | ID}.}

\item{n}{an integer number to specify the number of PDX lines.}

\item{m}{an integer number to specify the number of individuals per PDX line per treatment.}

\item{sim}{Number of Monte Carlo samples to be generated. Default is 1000.}

\item{two.sided}{A logical value to indicate if a two-sided hypothesis testing is conducted. Default is TRUE.}

\item{ncores}{number of cores for parallel computation.}
}
\description{
A function to obtain a power table with the combination of
various number of PDX lines and number of individuals per PDX lines per treatment based on a preliminary dataset
}
\examples{
require(PDXpower)
data(mice)

PowTab <- PowANOVADat(data = mice, formula = log(Y) ~ Tx, random = ~ 1|ID,
n = c(3, 5, 10), m = c(2, 3, 4))
PowTab
plotpower(PowTab[[4]], ylim = c(0, 1))

}
