% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerTable.R
\name{PowerTable}
\alias{PowerTable}
\title{A function to obtain a power table with the combination of
various number of PDX lines and number of individuals per PDX lines per treatment.}
\usage{
PowerTable(
  n,
  m,
  beta,
  tau2 = 0.5,
  alpha = 0.05,
  lambda = 0.03,
  nu = 2,
  sigma2 = 1,
  two.sided = TRUE,
  distr = c("Weibull", "normal"),
  lambdaC = 0.1,
  censor = TRUE,
  sim = 1000,
  print = c("both", "ANOVA", "Cox-frailty"),
  ncores = NULL
)
}
\arguments{
\item{n}{an integer number to specify the number of PDX lines.}

\item{m}{an integer number to specify the number of individuals per PDX line per treatment.}

\item{beta}{Treatment effect for the treated group.}

\item{tau2}{variance of PDOX line specific random effect.}

\item{alpha}{significance level. Default is 0.05.}

\item{lambda}{Scale parameter of Weibull distribution for the baseline hazard.}

\item{nu}{Shape parameter of Weibull distribution for the baseline hazard.}

\item{sigma2}{Error variance of log survival time for both treatment groups.}

\item{two.sided}{A logical value to indicate if a two-sided hypothesis testing is conducted.
Default is TRUE.}

\item{distr}{Distributional assumption of the simulated event time.}

\item{lambdaC}{Rate parameter of exponential distribution for the hazard of censoring.}

\item{censor}{logical value of whether a censoring distribution is considered in a data generation setting. Deafult is TRUE.}

\item{sim}{Number of Monte Carlo samples to be generated. Default is 1000.}

\item{print}{a string to indicate which model result to be printed.
If \code{print = "both"}, then the power curves of both models will be printed.
Otherwise, print a power curve from one of the two models by specifying
either \code{print = "ANOVA"} or \code{print = "Cox-frailty"}.}

\item{ncores}{number of cores for parallel computation.}
}
\description{
A function to obtain a power table with the combination of
various number of PDX lines and number of individuals per PDX lines per treatment.
}
\examples{

require(PDXpower)
data <- SimPDXdata(seed = 1000, n = 3, m = 3, beta = -0.8, tau2 = 0.2, lambda = 0.03,
nu = 2, sigma2 = 1, distr = "Weibull", lambdaC = 0.1, censor = TRUE)
CoxRandom <- frailtypack::frailtyPenal(survival::Surv(Y,status) ~ Tx + cluster(ID),
                                       data=data, RandDist = "LogN",
                                       print.times = FALSE, maxit = 50, hazard = "Weibull")

lambda <- CoxRandom$scale.weib[1]^(-CoxRandom$shape.weib[1])
nu <- CoxRandom$shape.weib[1]
beta <- CoxRandom$coef
tau2 <- CoxRandom$sigma2

n <- c(3, 5, 10)
m <- c(2, 3, 4)
fit <- PowerTable(n = n, m = m, beta = beta, lambda = lambda, nu = nu,
                  tau2 = tau2, distr = "Weibull", sim = 100,
                  censor = FALSE,
                  print = "both",
                  ncores = 7)

plotpower(fit, ylim = c(0.5, 1))

}
\seealso{
\code{\link{plotpower}}
}
