% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowFrailtyDat.R
\name{PowFrailtyDat}
\alias{PowFrailtyDat}
\title{A function to obtain a power table with the combination of
various number of PDX lines and number of individuals per PDX lines per treatment based on a preliminary dataset}
\usage{
PowFrailtyDat(
  data = NULL,
  formula = NULL,
  maxit = 50,
  hazard = "Weibull",
  n = NULL,
  m = NULL,
  sim = 1000,
  censor = FALSE,
  two.sided = TRUE,
  ncores = NULL
)
}
\arguments{
\item{data}{data.frame in which to interpret the variables named in the formula.}

\item{formula}{a formula object, with the response on the left of a ~ operator, and the terms on the right.
The response must be a survival object as returned by the Surv function.}

\item{maxit}{maximum number of iterations needed for model fitting. Default is 50.}

\item{hazard}{distributional assumption of the baseline hazard. Default is Weibull.}

\item{n}{an integer number to specify the number of PDX lines.}

\item{m}{an integer number to specify the number of individuals per PDX line per treatment.}

\item{sim}{Number of Monte Carlo samples to be generated. Default is 1000.}

\item{censor}{logical value of whether a censoring distribution is considered in a data generation setting. Default is FALSE.}

\item{two.sided}{A logical value to indicate if a two-sided hypothesis testing is conducted. Default is TRUE.}

\item{ncores}{number of cores for parallel computation.}
}
\description{
A function to obtain a power table with the combination of
various number of PDX lines and number of individuals per PDX lines per treatment based on a preliminary dataset
}
\examples{
require(PDXpower)
data <- SimPDXdata(seed = 1000, n = 3, m = 3, beta = -0.8, tau2 = 0.2, lambda = 0.03,
nu = 2, sigma2 = 1, distr = "Weibull", lambdaC = 0.1, censor = TRUE)

PowTab <- PowFrailtyDat(data = data, formula = Surv(Y,status) ~ Tx + cluster(ID),
n = c(3, 5, 10), m = c(2, 3, 4))
PowTab
plotpower(PowTab, ylim = c(0, 1))

}
